refactor(AddItem): MVC拆分主要逻辑
