+++ cs.HoLMod.AddItem/Modules/LanguageManager.cs
@@ -1203,70 +1203,10 @@ internal class LanguageManager
             {"邪龙", "Xielong"}
         };
     }
-
-    // 设置当前语言
-    public void SetLanguage(string languageCode)
-    {
-        if (translations.ContainsKey(languageCode))
-        {
-            currentLanguageCode = languageCode;
-        }
-        else
-        {
-            // 如果请求的语言不存在，使用默认语言
-            currentLanguageCode = "zh-CN";
-        }
-    }
-
-    // 获取翻译
-    public string GetTranslation(string key)
-    {
-        if (translations[currentLanguageCode].ContainsKey(key))
-        {
-            return translations[currentLanguageCode][key];
-        }
-        else
-        {
-            // 如果键不存在，返回键本身
-            return key;
-        }
-    }
-
-    // 获取支持的语言列表
-    public List<string> GetSupportedLanguages()
-    {
-        return translations.Keys.ToList();
-    }
-
-    // 获取语言显示名称
-    public string GetLanguageDisplayName(string languageCode)
-    {
-        switch (languageCode)
-        {
-            case "zh-CN": return "中文简体";
-            case "en-US": return "English";
-            // 可以在这里添加更多语言的显示名称
-            default:
-                return languageCode;
-        }
-    }
-    
-    // 检查系统语言是否为中文
-    public bool IsChineseLanguage()
-    {
-        // 检查系统语言是否为中文
-        string systemLanguage = Application.systemLanguage.ToString();
-        return systemLanguage.Contains("Chinese");
-    }
     
     // 根据当前语言获取文本
     public string GetText(string key)
     {
-        string languageCode = IsChineseLanguage() ? "zh-CN" : "en-US";
-        if (translations.ContainsKey(languageCode) && translations[languageCode].ContainsKey(key))
-        {
-            return translations[languageCode][key];
-        }
-        return key; // 如果没有找到对应的翻译，返回原文本
+        return "key"; // 如果没有找到对应的翻译，返回原文本
     }
 }
