+++ cs.HoLMod.AddItem/Views/BaseView.cs
@@ -1,0 +1,6 @@
+﻿namespace cs.HoLMod.AddItem;
+
+public class BaseView
+{
+    
+}
