+++ cs.HoLMod.AddItem/locales/en-US/AddItem.json
@@ -1,0 +1,33 @@
+﻿{
+  "Mod": {
+    "Name": "Item Adder",
+    "Author": ""
+  },
+  "Description": {
+    "1": "1. Please save the game before adding items for rollback",
+    "2": "2. Press F2 to show/hide window",
+    "3": "3. Switch modes: Items/Currency/Stories/Map",
+    "4": "4. Enter partial characters to search items or stories",
+    "5": "5. Select item, choose/enter quantity, then click Add"
+  },
+  "MenuTab": {
+    "Currency": "",
+    "Items": "",
+    "Map": "",
+    "Stories": ""
+  },
+  "PropCategory": {
+    "Other": "Other",
+    "Special": "Special"
+  },
+  "Info": {
+    "Search": "Search:",
+    "Category": "",
+    "Clear": "",
+    "Description": "",
+    "Count": ""
+  },
+  "Button": {
+    "Add": "Add"
+  }
+}
